Redefine BridgeGame interface
BridgeGame was not really as universally useful interface as its placement in
root Bridge namespace and its name would suggest. It's meant to interface the
control of "this" bridge instance (as opposed to an instance controlled by
other players) over the game. Renamed BridgeGame to BridgeController and moved
it to Bridge::Main namespace to better reflect the use of the interface.