Make Bridge application support one client per player
This commit effectively turns the backend a server which supports a signle
session with four players each controlling one position. Command are now
accepted from client for one player only. Also the backend publishes each client
only deal state for its player.

The player each client controls is determined by the order they join the
game. The first player is given north seat, the second east and so on. Leaving
and rejoining the game is not yet handled. The backend just rejects any client
that tries to join after four clients have connected.

Later backend-to-backend protocol will be added to enable peer-to-peer games
in addition to traditional client-server model.