Make getWinner() nonmember function
Previously getWinner() was member of Trick that delegated to virtual
implementation handleGetWinned(). However, rules of determining the winner of
the trick should not depend on the specific implementation of Trick. The current
public interface of Trick was enough to implement the algorithm, so it was moved
from virtual member of BasicTrick to nonmember function.

BasicTrick was changed not to take trump as constructor argument as trump is now
argument of getWinner().