Move determining allowed cards to trick to the base class
Previously Trick delegated the decision whether or not a card was allowed to the
trick according to the bridge rules to its derived class using virtual
handleIsPlayAllowed call. However, as the rule does not depend on the
implementation of the trick object, the algorithm was moved to the base class.