Start developing peer protocol
This commit includes

- Adding PeerClientControl to keep track of peers and clients of the
  application
- Taking PeerClientControl into use in BridgeMain (although it doesn't yet
  allow actual peers to join)
- Adding position argument to commands of the Bridge protocol

The last change is necessary to be able to make peers control several
players. Identification of the player for which the command is issued makes
the protocol more robust and less suspectible to state mismatched. For
uniformity the identification is also required from clients although clients
only control one player.