Resend message to peers if the reply is failure
This commit introduces class PeerCommandSender which queues messages and only
sends them if all peers have replied success to the previous messages. If a
peer replies failure, the message is resent. The assumption is that failures
are due to temporary discrepancies in states of the peers and will be
synchronized again when all peers have handled all previous commands from
other peers.

Current solution is still not perfect as misbehaving peer can cause infinite
resend loops that is currently not terminated.