Overload swap for TMCG_Stack
In TMCG::shuffle() method swapping of stacks was performed but in practice this
was not useful as TMCG_Stack is not movable nor was swap implemented for
it. Having custom swap makes swapping more efficient.

Because LibTMCG API documentation includes (the public) stack member variables,
it can be reasonably said that swapping the internal representation does not
break encapsulation.