Refactor BridgeMain to use SimpleCardProtocol instead of implementing it
This commit takes SimpleCardProtocol into use in BridgeMain, contrary to the
earlier desing where BridgeMain implemented the protocol directly. This design
allows pluggable protocol to be used in the future.

Also a bug was noticed that made it possible for deal and play commands to be
sent out of order if the leader played the last card to the deal.