Also publish hand and card when publishing play event
Hand position and type of the played card are added as arguments to the play
event published when card is played to the trick.

The new information is used by the GUI to ensure that full trick is shown
after completion, making it possible for all the players to examine the fourth
card played to the trick before a card is lead to the next trick.