Remove command strings from BridgeMain.hh and CardServerMain.hh
Command strings are not part of the public class interface in the sense that
they are not used as parameters when calling any functions of BridgeMain and
CardServerMain. Rather they are needed when sending commands to the
application. This happens typically in a separate process where the strings
are not accessible.