Fix early destruction of std::shared_ptr<Hand> in BridgeEngine
In BridgeEngine cards are played when CardRevealState::COMPLETED notification is
received from the hand. However, after playing the last card of the deal, the
state machine immediately moves to Shuffling state and (without this fix) would
destroy the std::shared_ptr<Hand> objects stored by InDeal state - while still
handling the notification.

To work around the problem, a temporary handle is added to BridgeEngine::Impl
Assigning the pointers to this variable artificially expands their lifetime so
that the notification can be completed safely. Before the next shuffle is
requested (and thus the hand invalidated), the handle is released.