Add CardServerProxy
CardServerProxy is the client for card server inside the main backend
application. It is responsible for communicating with the card server and
providing the CardProtocol style interface to the cards.

This change also includes means to configure the backend application to
communicate with the card serve. With this change it is possible to establish
secure (in terms of card exchange) mental card game bridge session.