Use Mersenne twister in SimpleCardProtocol
Previously std::default_random_engine was used when generating random numbers in
SimpleCardProtocol. At least with GCC this is the "minimal standard"
minstd_rand0 engine. As the C++ standard library readily implements the Mersenne
twister algorithm which offers better quality primes, it is used instead of the
default rangom engine.