Do not return bool from message handlers in MessageLoop
Previously there were two ways to terminate message loop: by explicit call to
MessageLoop::terminate() and by returning false from a message handler. To keep
the interface cleaner, the call to terminate() is now the only way to terminate
the loop. The latter way was never used in the program and is just additional
complexity.