Initial commit for proper GUI
The GUI included earlier was crude test tool with various shortcomings, most
notably lack of virtually any code structure and error checking.

The new GUI is written in PyQt5, as opposed to the earlier GUI written in
Kivy. Kivy appears to be targeted for more demanding games, not more tradition
lightweight widget based GUIs. In addition programs written in PyQt5 seem to
support easier to deploy.