Start structuring Bridge GUI into packages and modules
Bridge GUI is full-fledged program written in Python that will eventually need
well structured modules and unit testing. This change introduces the first step
towards the goal by moving freeze() utility into a separate module and writing
unit test for it.

Unit testing is not yet part of the build system but eventually it is preferable
to have it so.