Add position utilities to positions module
Positions, unlike cards and calls, have mostly been handled in the bridge GUI
using the serialized representation. This is somewhat error prone as it puts the
burden of handling deserialization errors in the code dealing with what could
already be internal representation of the position.

This change introduces bridgegui.positions.asPosition() which can be used to do
the conversion in one well-defined point and use the internal representation (an
enumeration) in other modules. Other utilities have been implemented in terms of
the internal representation.