Create setup.py script for bridgegui
Beginning from this commit, setuptools can be used to build the bridgegui script
and package required to run the GUI. The images are included as data files for
the bridgegui package, exploiting the pkg_resources API to find them (instead of
an environment variable).

In addition the setup is integrated into the CMake build system. This means that
the 'make' phase invokes 'python setup.py build', the 'make test' phase invokes
'python setup.py test' and the 'make install' phase invokes 'python setup.py
install'.