Define executable output path in CMake files
Previously the name of the main output executable (bridge) clashed with the main
source directory (also bridge) which lead to problems in the build. This was
previously worked around by naming the source directory "bridge_". Now that the
executable output path is defined, the name clash is no longer issue so the
underscore from the source directory could be removed.