Clarify bridge protocol specification
The goal of the clarifications is especially with consistent terminology: The
vague term backend was replaced with more precise term (bridge) peer. Peers and
clients are now collectively known as nodes, and their distinct responsibilities
are now stated more explicitly.

Attempt was also made to clarify the card server protocol specification and to
use the new terminology. However, it is acknowledged that more clarification is
necessary.