Introduce game command to the bridge protocol
This commit introduces the first step towards supporting multiple games in
single peer, by conceptually splitting the handshake command (bridgehlo) and the
command for joining a game (game). In addition the role (client or peer) and the
version of the protocol are now established in the handshake. Note that there is
still only one set of state machines for handling a game, thus multiple games
are not actually supported.

This change breaks card server protocol, which in this commit is commented out
rather than fixed.