Fix compiler errors reported by GCC 6.3.0
The compiler errors are
- Missing include in NodeControl.hh
- std::make_pair not working correctly when constructing reference to incomplete
  type in Trick.hh

I'm not sure if the latter is actually bug in my libstdc++ because there seems
to be no reason why the reference held by the pair should be to a complete
type. However, with the fix being clean I decided to do it anyway and
investigate the std::make_pair issue later.