Send status code in replies as binary
Previously strings "success" and "failure" were used to indicate successful or
failed reply, respectively. Using status code instead gives more flexibility to
communicate reason for the success/failure.

Although parameters to the commands are normally serialized as JSON, the status
code has binary representation. This is because the meaning of the status code
is fixed so nothing as flexible as JSON is needed to interpret it. Note that the
old statuses (success/failure) were not serialized either.