Move handling of call() and play() commands to BridgeGame
This commit turns call() and play() in BridgeMain to wrappers whose handling is
done in BridgeGame class.

Currently PeerCommandSender is needed both in BridgeMain and BridgeGame, which
is solved for now by pushing the definitions of auxiliary methods to the header
of BridgeGame. This should be solved better in the following commits by
separating better the node handling roles in BridgeMain (should maintain list of
all peers and nodes the backend talks to) and BridgeGame (sends commands to only
those peers that are taking part in the game).