Initialize BridgeGame in single stage
This commit removes the two-way dependency which prevented BridgeGame from
being initialized in one stage. This was achieved by giving BridgeMain the
responsibility of setting up the card protocol and relationship between peers,
and passing that information to BridgeGame as parameters of type CardProtocol
and PeerCommandSender.

BridgeMain now deals with configuring the game based on arguments provided by
the invoker of the program, and BridgeGame deals with what's necessary to
actually run the game after configuration.