Allow clients to create game
This commit implements game command for clients. The command creates new,
peerless game where the backend instance itself controls all positions and to
which clients are able to join.

Setting up the game is currently quite cumbersome, because it involves creating
peer communication facilities not really needed for anything. In the future more
lightweight solution could be deviced for peerless games.