Select game for client if not explicitly specified
This commit introduces a queue of joinable games in BridgeMain. If a client
joins without specifying a game, select game from that queue if possible.

Limitations: Do not try to select game if client has also specified
position (would complicate popping from the queue otherwise). Also if there are
no games in the queue, try joining the null game present in peer mode.