PeerCommandSender: delay resending message on failed reply
Previously if a peer replied failure, the command was immediately resent to the
peer with the assumption that the peer would promptly catch up with the state of
the other peers and accept the message. Because there was no delay, this could
lead to tight loop of failed replies and resending the message.

This commit introduces delay with increasingly long intervals between receiving
the failed reply and attempting to resend the message.