Introduce config files and configurable curve keys
This commit adds several features and interfaces supporting configuring curve
keys used for communication.

The core addition is Bridge::Main::Config class which acts as configuration file
parser. The config files are Lua scripts. Dependency to Lua library is
introduced for processing the scripts.

setupCurveServer() and setupCurveClient() now require keys as arguments. Keys
are introduces in most places in the code to enable the new interface.