Block SIGINT and SIGTERM in worker threads
There was a bug due to CallbackScheduler creating worker thread which did not
set signal mask. Due to this it would handle SIGINT and SIGTERM by terminating
the program, not allowing the main thread to exit cleanly.

This commit adds Bridge::Thread class for handling the signal blocking and
joining of the thread.