Make MessageHandler byte oriented
MessageHandler is clearly an interface that should be agnostic about the
underlying encoding of the messages, so Blob is more suitable datatype for its
interface than the std::string (which has UTF-8 encoding implied).

This commit most importantly turns MessageHandler into byte oriented interface,
and implements necessary changes that follow from it, most notably refactoring
FunctionMessageHandler. The refactoring is not complete and some message
handlers are still string oriented with minimal changes to support the new
interface.

Other changes

- Implement classes supporting non-copying byte oriented interface (ByteSpan and
  BlobMap)

- Avoid copies where necessary by choosing between Blob and ByteSpan, where
  resource allocating std::string was previously used by default.

- Reimplement MessageQueue to use the new MessageHandler interface, and also to
  avoid making copies by directly passing views to the ZeroMQ messages to the
  handler.

- Delete old functions that became unused in the process.

- Clarify documentation concerning bytes vs. characters