Send CURVE server key to card server
This commit adds CURVE server key as field in the serialized PeerEntry sent to
card server as part of the initialization, and game command the peers send to
each other.

The card server public key is assumed to be the same as the bridge backend
public key which simplifies the implementation somewhat. This restriction can be
lifted in the future if necessary.