Split Identity into user and routing IDs
This commit turns Messaging::Identity into struct containing userId and
routingId. userId is intended to be used by the application to strongly
authenticate a node.

Definition of the Identity property also requires changes to code dealing with
Identity object. Where identity was used as opaque object (as map key for
instance) changes are minimal, but notably CardServerProxy used serialized
identity as way to identify a peer. Identity::routingId is used in this case as
the key.