Simplify MessageLoop implementation
This commit simplifies the logic of MessageLoop by making callbacks and
pollitems vectors same length.

Previously there was one extra pollitem for the signal handler whose handling
was hardcoded in the polling loop in Impl::run(). Now the signal handler is also
implemented as callback, making the polling loop oblivious to what kind of items
are being processed.

In order to make different kinds of callbacks (socket and fd) the callbacks
vector now contains variants of socket and fd callbacks, and the calls are
dispatched by a visitor.