Parameterize MessageHandler with ExecutionPolicy
This commit adds a pluggable execution policy to MessageHandler, making it a
template. This is meant to implement support for execution out of MessageQueue's
call stack, i.e. in coroutines or even in other threads.

No actually interesting execution policies are introduced in this commit, and
the only policy is synchronous policy executing as it used to.