Add ensureSocketReadable()
ensureSocketReadable() is a (group of) function (overloads) that can be used
to... ensure that a socket is readable. The idea is that for synchronous
execution policy it simply polls, but for asynchronous execution it suspends the
coroutine to await the socket.

Using the extra level of indirection allows writing generic code for different
execution policies that need to poll/await a socket.