card server: make message handlers asynchronous
So far CardServerMain has defined message handlers that execute synchronously,
meaning that a command blocks until it has executed the whole
protocol (shuffling, revealing, drawing etc.) with its peers. This makes the
process not responsive while a protocol is being executed.

By making the message handlers coroutines that suspend execution when they are
waiting messages from card server peers, control returns to the main event loop
periodically and other messages / signals can be handled normally.

Note! It is not yet possible to remove the terminate-immediately-on-signal
behavior because coroutines cannot be canceled cleanly. All sorts of nasty
things happen if the coroutine is destructed at the very late stage of the
program (when the shared_ptr help by MessageLoop goes out of scope) and
dependencies have been destructed in the meantime. Controlled unwinding of
coroutine stack is needed.