Make CoroutineAdapter store Poller and CallbackScheduler by weak_ptr
This makes it more robust if Poller and CallbackScheduler go out of scope
because of shutting down the program while a coroutine is running. weak_ptr is
used instead of shared_ptr because Poller and CallbackScheduler are required to
capture the coroutine by shared_ptr, so breaking of cycle is needed.