Remove SignalGuard from CardServerMain
SignalGuard was originally used to unset signal block mask while running TMCG
protocols, which would potentially cause the execution to block
indefinitely. Now that the control is returned to the message loop while waiting
for messages (thanks to the coroutine implementation), this is
unnecessary. Signal can be handled synchronously and the program terminated
normally.