PollingCallbackScheduler: send commands without waiting
When the working thread is terminated, but before the PollingCallbackScheduler
object is destructed, send on PAIR socket will block in case a callback is
scheduled. Use ZMQ_DONTWAIT flag to avoid this situation. Failure to send the
message is ignored because it is assumed the program is in process of being
terminated anyway.