Define cppzmq classes/methods in Sockets.hh header
The bridge messaging framework is tightly coupled with the cppzmq library,
including using the bare types/methods from the library in upper layers. In
order to abstract the dependency a little, header messaging/Sockets.hh was added
to contain typedefs for the types introduced by cppzmq, as well as small
wrappers for the commonly used methods.