Replace integer statuses with strings
The bridge protocol doesn't use binary encoded integers outside of the
status frame. In fact the protocol uses JSON serialized integers
whenever it needs integers in payload, which makes two highly
different approaches to encoding integers in the same protocol.

For consistency with the rest of the protocol definition, use status
frames OK and ERR to indicate successful and failed replies.