Always create unknown players
Previously NodePlayerControl class had separate methods to get player
and create a new player. The join command would create player and the
other commands would just retrieve an existing player.

To simplify the behavior and ensure that any node can claim a player
by issuing a command referring to it, this commit replaces the
separate createPlayer() and getPlayer() methods with a
getOrCreatePlayer() method which is used everywhere where a command
handler needs to do access control for a player.