Refactoring: Extract deal state information to Deal object
After this commit, a big chunk of deal state information directly
embedded in BridgeEngine is exposed via a newly created Deal
interface.

This isn't much itself, but will later enable Deal objects to exist
independently of a BridgeEngine. This makes it simpler to persist
deals for records and state recovery.

As a minor modification to the protocol, get reply now returns
null objects for deal state if the deal does not yet exist (instead of
an object with lots of dummy data).