Make player argument mandatory in the protocol
Until now it was ok for clients controlling a single player to omit the player
argument in commands. This adds unnecessary complexity, and ties players to
connections tightly. In order to simplify the implementation, and decouple
connections and players, add player as a mandatory argument.

This commit also changes the NodePlayerControl to authenticate based on UserId
only. "Authentication" based on routing ID was only enabling omitting player for
connections only controlling single players, and didn't actually have anything
to do with security.